@@ -4,6 +4,4 @@ declare(strict_types=1);
 
 namespace Derhansen\FeChangePwd\Exception;
 
-use Exception;
-
-class MissingFeatureToggleException extends Exception {}
+class MissingFeatureToggleException extends \Exception {}
