@@ -12,44 +12,37 @@ declare(strict_types=1);
 namespace Derhansen\FeChangePwd\Service;
 
 use Psr\Http\Message\ServerRequestInterface;
-use TYPO3\CMS\Core\Context\Context;
-use TYPO3\CMS\Core\Context\TypoScriptAspect;
+use TYPO3\CMS\Core\Site\Entity\Site;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
- * Service with helper function for settings handling
+ * Service with helper function for handling typoscript settings
  */
 class SettingsService
 {
     protected array $settings = [];
 
     /**
-     * Returns the settings
+     * Returns TypoScript settings
      */
-    public function getSettings(ServerRequestInterface $request): array
+    public function getTypoScriptSettings(ServerRequestInterface $request): array
     {
         if (empty($this->settings)) {
-            try {
-                $fullTypoScript = $request->getAttribute('frontend.typoscript')->getSetupArray();
-            } catch (\Exception $e) {
-                // An exception is thrown, when TypoScript setup array is not available. This is usually the case,
-                // when the current page request is cached. Therefore, the TSFE TypoScript parsing is forced here.
-
-                // Set a TypoScriptAspect which forces template parsing
-                GeneralUtility::makeInstance(Context::class)
-                    ->setAspect('typoscript', GeneralUtility::makeInstance(TypoScriptAspect::class, true));
-
-                // Call TSFE getFromCache, which re-processes TypoScript respecting $forcedTemplateParsing property
-                // from TypoScriptAspect
-                $tsfe = $request->getAttribute('frontend.controller');
-                $requestWithFullTypoScript = $tsfe->getFromCache($request);
-
-                $fullTypoScript = $requestWithFullTypoScript->getAttribute('frontend.typoscript')->getSetupArray();
-            }
+            $fullTypoScript = $request->getAttribute('frontend.typoscript')->getSetupArray();
             $settings = $fullTypoScript['plugin.']['tx_fechangepwd.']['settings.'] ?? [];
             $this->settings = GeneralUtility::removeDotsFromTS($settings);
         }
         return $this->settings;
+    }
+
+    /**
+     * Returns site settings for the extension
+     */
+    public function getSiteSettings(ServerRequestInterface $request): array
+    {
+        /** @var Site $site */
+        $site = $request->getAttribute('site');
+        return $site->getSettings()->get('fe_change_pwd');
     }
 
     /**
